+++ lua/plugins/config/flutter.lua
@@ -1,88 +1,62 @@
 return function(use)
 	use({
 		'akinsho/flutter-tools.nvim',
-		wants = { 'plenary.nvim', 'nvim-lspconfig', 'telescope.nvim', 'nvim-dap' },
+		requires = {
+			'nvim-telescope/telescope.nvim',
+			'neovim/nvim-lspconfig',
+			'mfussenegger/nvim-dap',
+			'nvim-lua/plenary.nvim',
+		},
 		config = function()
 			local ok, telescope = pcall(require, 'telescope')
 			if ok then
 				telescope.load_extension('flutter')
 			end
 
-			local lsp_installer_servers = require('nvim-lsp-installer.servers')
-			local server_available, dartls = lsp_installer_servers.get_server('dartls')
-			if server_available then
-				dartls:on_ready(function()
-					require('flutter-tools').setup({
-						debugger = {
-							enabled = true,
-							run_via_dap = false,
-						},
-						widget_guides = {
-							enabled = true,
-						},
-						lsp = {
-							on_attach = function(client, bufnr)
-								require('plugins.config.lsp.on_attach')(client, bufnr)
-								local wk = require('which-key')
-								wk.register({
-									['<leader>g'] = {
-										name = '+flutter',
-										c = {
-											'<cmd>Telescope flutter commands<cr>',
-											'Commands',
-										},
-										s = {
-											'<cmd>FlutterRun<cr>',
-											'Run',
-										},
-										e = {
-											'<cmd>FlutterEmulators<cr>',
-											'Emulators',
-										},
-										r = {
-											'<cmd>FlutterReload<cr>',
-											'Reload',
-										},
-										R = {
-											'<cmd>FlutterRestart<cr>',
-											'Restart',
-										},
-										d = {
-											'<cmd>FlutterDevices<cr>',
-											'Devices',
-										},
-										q = {
-											'<cmd>FlutterDevices<cr>',
-											'Quit',
-										},
-										t = {
-											'<cmd>FlutterDevTools<cr>',
-											'Dev Tools',
-										},
-										y = {
-											'<cmd>FlutterCopyProfileUrl<cr>',
-											'Copy Profile URL',
-										},
-										b = {
-											'<cmr>FlutterLspRestart<cr>',
-											'LSP Restart',
-										},
-									},
-									gs = { '<cmd>FlutterOutlineToggle<cr>', 'Widget Outline' },
-								}, { buffer = bufnr })
-							end,
-							capabilities = require('plugins.config.lsp.capabilities'), -- e.g. lsp_status capabilities
-							settings = {
-								showTodos = true,
-								completeFunctionCalls = true,
-							},
-						},
-					})
-				end)
-				if not dartls:is_installed() then
-					dartls:install()
-				end
-			end
+			require('flutter-tools').setup({
+				debugger = {
+					enabled = true,
+					run_via_dap = false,
+				},
+				widget_guides = {
+					enabled = true,
+				},
+				lsp = {
+					on_attach = function(client, bufnr)
+						require('plugins.config.lsp.on_attach')(client, bufnr)
+
+						local function map(key, command, desc)
+							vim.keymap.set(
+								'n',
+								'<leader>g' .. key,
+								command,
+								{ desc = 'Flutter ' .. desc, buffer = bufnr }
+							)
+						end
+						map('c', '<cmd>Telescope flutter commands<cr>', 'Commands')
+						map('s', '<cmd>FlutterRun<cr>', 'Run')
+						map('e', '<cmd>FlutterEmulators<cr>', 'Emulators')
+						map('r', '<cmd>FlutterReload<cr>', 'Reload')
+						map('R', '<cmd>FlutterRestart<cr>', 'Restart')
+						map('d', '<cmd>FlutterDevices<cr>', 'Devices')
+						map('q', '<cmd>FlutterQuit<cr>', 'Quit')
+						map('t', '<cmd>FlutterDevTools<cr>', 'Dev Tools')
+						map('y', '<cmd>FlutterCopyProfileUrl<cr>', 'Copy Profile URL')
+						map('b', '<cmd>FlutterLspRestart', 'LSP Restart')
+						vim.keymap.set(
+							'n',
+							'gs',
+							'<cmd>FlutterOutlineToggle<cr>',
+							{ desc = 'Flutter Widget Outline', buffer = bufnr }
+						)
+					end,
+					capabilities = require('plugins.config.lsp.capabilities'), -- e.g. lsp_status capabilities
+					settings = {
+						showTodos = true,
+						completeFunctionCalls = true,
+					},
+				},
+			})
 		end,
 		ft = { 'dart' },
 	})