+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - **lsp:** rust tools move to top
 - **packer:** added check if ls failed
 
+> Various Actions
+- **snapshot:** update
+
 
 <a name="v4.7.0"></a>
 ### [v4.7.0] - 2022-05-19