+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 > Features
 - **lsp:** rust tools move to top
 - **packer:** added check if ls failed
+- **telescope:** mappings changed from using which-key to vim.keymap.set
 
 > Various Actions
 - **snapshot:** update snapshot