+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <a name="latest"></a>
 ### [Latest]
 
+> Code Refactoring
+- **flutter:** refactor flutter tools plugin config
+
 > Features
 - **lsp:** rust tools move to top
 - **packer:** added check if ls failed