+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 <a name="latest"></a>
 ### [Latest]
 
+> Features
+- **fzf-lua:** added description to mappings
+
+
+<a name="v4.8.0"></a>
+### [v4.8.0] - 2022-05-23
+
 > Code Refactoring
 - **flutter:** removed pcall
 - **flutter:** refactor flutter tools plugin config
@@ -1119,7 +1126,8 @@
 - initial commit
 
 
-[Latest]: https://github.com/tigorlazuardi/nvim/compare/v4.7.0...HEAD
+[Latest]: https://github.com/tigorlazuardi/nvim/compare/v4.8.0...HEAD
+[v4.8.0]: https://github.com/tigorlazuardi/nvim/compare/v4.7.0...v4.8.0
 [v4.7.0]: https://github.com/tigorlazuardi/nvim/compare/v4.6.1...v4.7.0
 [v4.6.1]: https://github.com/tigorlazuardi/nvim/compare/v4.6.0...v4.6.1
 [v4.6.0]: https://github.com/tigorlazuardi/nvim/compare/v4.5.3...v4.6.0